messaging_app/requirements.txt
asgiref==3.8.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.1
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-nested-routers==0.94.2
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
mysqlclient==2.2.7
promise==2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.1
urllib3==2.5.0
